requirements.txt
# requirements.txt
Django==5.2
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8
django-extensions==3.2.3
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-filter==23.5
django-import-export==3.3.5
celery==5.3.4
redis==5.0.1
PyJWT==2.8.0
pytz==2023.3
plotly==5.18.0
pandas==2.1.4
numpy==1.26.2
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
# boto3==1.34.14  # For AWS S3
django-storages==1.14.2
# stripe==7.8.0  # For payments (future)
# sendgrid==6.11.0  # For email
# twilio==8.11.0  # For SMS